requirements.txt
geopandas
nodejs
geoserver-rest
zip